feat: Add temporary headers configuration to allow mixed content for testing
This commit introduces a temporary headers configuration in next.config.ts to allow mixed content (HTTP API from an HTTPS site) during testing. A Content-Security-Policy header is added to upgrade insecure requests. This change is intended for testing purposes and will be removed once HTTPS is set up on the backend.